+++ src/translator/libxml_archive_read.cpp
@@ -323,7 +323,7 @@ int archiveRead(void* context, char* buffer, int len) {
     if ((signed)size < 0)
         return 0;
 
-    return size;
+    return (int)size;
 }
 
 void archiveDeleteContext(void* context) {
